refactor: Unify file naming and context propagation test conventions in internal/domain
