+++ internal/domain/repository/service_test.go
@@ -194,12 +194,12 @@ func Test_contextPropagation(t *testing.T) {
 		name string
 		call func(t *testing.T, m *core.Method)
 	}{
-		{"RepositoryService.List", func(t *testing.T, m *core.Method) {
+		{"List", func(t *testing.T, m *core.Method) {
 			m.Get = makeMockFn(t)
 			s := repository.NewService(m)
 			s.List(ctx, testProject) //nolint:errcheck
 		}},
-		{"RepositoryService.One", func(t *testing.T, m *core.Method) {
+		{"One", func(t *testing.T, m *core.Method) {
 			m.Get = makeMockFn(t)
 			s := repository.NewService(m)
 			s.One(ctx, testProject, testRepo) //nolint:errcheck
