+++ internal/domain/space/activity_service.go
