+++ internal/domain/space/attachment_service.go
