+++ internal/domain/user/star_service.go
