+++ internal/domain/user/star_service_test.go
