+++ internal/domain/wiki/star_service.go
