+++ internal/domain/wiki/star_service_test.go
